+++ internal/oci/distribution.go
@@ -1,10 +1,9 @@
-package registry
+package oci
 
 import (
-	"fmt"
 	"regexp"
 
-	"github.com/containerd/containerd/reference"
+	"github.com/opencontainers/go-digest"
 )
 
 // Package is used to parse components from requests which comform with the OCI distribution spec.
@@ -18,81 +17,72 @@ var (
 	blobsRegexDigest    = regexp.MustCompile(`/v2/` + nameRegex.String() + `/blobs/(.*)`)
 )
 
-func manifestWithTagReference(registry, path string) (reference.Spec, bool, error) {
+func manifestWithTagReference(registry, path string) (Image, bool, error) {
 	comps := manifestRegexTag.FindStringSubmatch(path)
 	if len(comps) != 6 {
-		return reference.Spec{}, false, nil
+		return Image{}, false, nil
 	}
-	ref, err := reference.Parse(fmt.Sprintf("%s/%s:%s", registry, comps[1], comps[5]))
-	if err != nil {
-		return reference.Spec{}, true, err
-	}
-	return ref, true, nil
+	img := NewImageWithTag(registry, comps[1], comps[5])
+	return img, true, nil
 }
 
-func manifestWithDigestReference(registry, path string) (reference.Spec, bool, error) {
+func manifestWithDigestReference(registry, path string) (Image, bool, error) {
 	comps := manifestRegexDigest.FindStringSubmatch(path)
 	if len(comps) != 6 {
-		return reference.Spec{}, false, nil
+		return Image{}, false, nil
 	}
-	ref, err := reference.Parse(fmt.Sprintf("%s/%s@%s", registry, comps[1], comps[5]))
-	if err != nil {
-		return reference.Spec{}, true, err
-	}
-	return ref, true, nil
+	img := NewImageWithDigest(registry, comps[1], digest.Digest(comps[5]))
+	return img, true, nil
 }
 
 // ManifestReference parses name and reference components from manifest path and returns an image reference.
 // If path does not match any of the regex patterns false will be returned without an error.
 // /v2/<name>/manifests/<reference>
-func ManifestReference(registry, path string) (reference.Spec, bool, error) {
-	ref, ok, err := manifestWithTagReference(registry, path)
+func ManifestReference(registry, path string) (Image, bool, error) {
+	img, ok, err := manifestWithTagReference(registry, path)
 	if err != nil {
-		return reference.Spec{}, ok, err
+		return Image{}, ok, err
 	}
 	if ok {
-		return ref, ok, nil
+		return img, ok, nil
 	}
-	ref, ok, err = manifestWithDigestReference(registry, path)
+	img, ok, err = manifestWithDigestReference(registry, path)
 	if err != nil {
-		return reference.Spec{}, ok, err
+		return Image{}, ok, err
 	}
 	if ok {
-		return ref, ok, nil
+		return img, ok, nil
 	}
-	return reference.Spec{}, false, nil
+	return Image{}, false, nil
 }
 
 // BlobReference parses name and reference components from blob path and returns and image reference.
 // If path does not match the regex pattern false will be returned without an error.
 // /v2/<name>/blobs/<reference>
-func BlobReference(registry, path string) (reference.Spec, bool, error) {
+func BlobReference(registry, path string) (Image, bool, error) {
 	comps := blobsRegexDigest.FindStringSubmatch(path)
 	if len(comps) != 6 {
-		return reference.Spec{}, false, nil
+		return Image{}, false, nil
 	}
-	ref, err := reference.Parse(fmt.Sprintf("%s/%s@%s", registry, comps[1], comps[5]))
-	if err != nil {
-		return reference.Spec{}, true, err
-	}
-	return ref, true, nil
+	img := NewImageWithDigest(registry, comps[1], digest.Digest(comps[5]))
+	return img, true, nil
 }
 
 // Any reference returns the name and tag or digest for a path whcih matches any of the request paths.
-func AnyReference(registry, path string) (reference.Spec, bool, error) {
-	ref, ok, err := ManifestReference(registry, path)
+func AnyReference(registry, path string) (Image, bool, error) {
+	img, ok, err := ManifestReference(registry, path)
 	if err != nil {
-		return reference.Spec{}, ok, err
+		return Image{}, ok, err
 	}
 	if ok {
-		return ref, ok, nil
+		return img, ok, nil
 	}
-	ref, ok, err = BlobReference(registry, path)
+	img, ok, err = BlobReference(registry, path)
 	if err != nil {
-		return reference.Spec{}, ok, err
+		return Image{}, ok, err
 	}
 	if ok {
-		return ref, ok, nil
+		return img, ok, nil
 	}
-	return reference.Spec{}, false, nil
+	return Image{}, false, nil
 }